Creating examples

+++ contrib/node/src/python/pants/contrib/node/tasks/npm_resolve.py
@@ -19,23 +19,29 @@ from pants.util.dirutil import safe_mkdir
 from pants.contrib.node.tasks.node_task import NodeTask
 
 
-def _safe_copy(source, dest):
+def _safe_copy(source, dest, force_copy=False):
   safe_mkdir(os.path.dirname(dest))
-  try:
-    os.link(source, dest)
-  except OSError as e:
-    if e.errno == errno.EXDEV:
-      # We can't hard link across devices, fall back on copying
-      shutil.copyfile(source, dest)
-    else:
-      raise
+  if force_copy:
+    shutil.copyfile(source, dest)
+  else:
+    try:
+      os.link(source, dest)
+    except OSError as e:
+      if e.errno == errno.EXDEV:
+        # We can't hard link across devices, fall back on copying
+        shutil.copyfile(source, dest)
+      else:
+        raise
 
 
 def _copy_sources(buildroot, node_module, dest_dir):
   source_relative_to = node_module.address.spec_path
   for source in node_module.sources_relative_to_buildroot():
-    dest = os.path.join(dest_dir, os.path.relpath(source, source_relative_to))
-    _safe_copy(os.path.join(buildroot, source), dest)
+    source_relative_to_node_module_path = os.path.relpath(source, source_relative_to)
+    _safe_copy(
+      os.path.join(buildroot, source),
+      os.path.join(dest_dir, source_relative_to_node_module_path),
+      source_relative_to_node_module_path == 'package.json')
 
 
 class NpmResolve(NodeTask):
@@ -131,10 +137,31 @@ class NpmResolve(NodeTask):
       return node_paths.node_path(target) if self.is_node_module(target) else target.version
     dependencies = {dep.package_name: render_dep(dep) for dep in npm_package.dependencies}
 
-    package = {
-      'name': npm_package.package_name,
-      'version': '0.0.0',
-      'dependencies': dependencies
-    }
-    with open(os.path.join(node_path, 'package.json'), 'wb') as fp:
+    package_json_path = os.path.join(node_path, 'package.json')
+
+    if os.path.isfile(package_json_path):
+      with open(package_json_path, 'r') as fp:
+        package = json.load(fp)
+    else:
+      package = {}
+
+    if not package.has_key('name'):
+      package['name'] = npm_package.package_name
+    elif package['name'] != npm_package.package_name:
+      raise TaskError('Name in package.json not the same as the BUILD target name for {}'
+                      .format(npm_package.address.reference()))
+
+    if not package.has_key('version'):
+      package['version'] = '0.0.0'
+    elif package['version'] != npm_package.version:
+      raise TaskError('Version in package.json not the same as the BUILD target name for {}'
+                      .format(npm_package.address.reference()))
+
+    if not package.has_key('dependencies'):
+      package['dependencies'] = dependencies
+    else:
+      raise TaskError('package.json for {} lists dependencies. Dependencies must be specified in the BUILD file.'
+                      .format(npm_package.address.reference()))
+
+    with open(package_json_path, 'wb') as fp:
       json.dump(package, fp, indent=2)
